pyproject.toml
[project]
name = "color-harmony"
version = "0.1.0"
description = "Hue bucket explorer built with Gradio"
authors = [
    { name = "ankit K", email = "ankit48365@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "gradio>=6.13.0",
    "matplotlib>=3.10.9",
    "numpy>=2.4.4",
    "pillow>=12.2.0",
]

[project.scripts]
color-harmony = "app:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
only-include = ["src"]
sources = ["src"]

[dependency-groups]
dev = [
    "pytest>=9.0.3",
]
